Added optional arguments for cleanup and archive

+++ fundamentals_screener/file_system_ops/image_cleanup.py
@@ -1,7 +1,7 @@
 from glob import glob
 import os
 
-def image_cleanup():
+def image_cleanup(prompt_each=False):
     # list of all image files, if any
     image_file_paths = glob("C:\\Users\\andre\\PythonProjects\\fundamentals_screener\\*.png")
 
@@ -12,6 +12,12 @@ def image_cleanup():
         # removes each image file one by one, since wildcards do not work with os.remove()
         for image in image_file_paths:
             try:
-                os.remove(image)
+                # prompt each is set to true if '-p' flag is passed to 'cleanup'
+                if prompt_each:
+                    confirm = input("Remove " + image + "? [Y/N] ").upper()
+                    if confirm == 'Y':
+                        os.remove(image)
+                else:
+                    os.remove(image)
             except:
-                print("Error while removing file: " + image)
+                print("Error while removing file " + image)
